python_helper_bot/requirements.txt
aiogram==3.*
python-dotenv
pytest